pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
authors = ["Eugene <eugeneturkin@gmail.com>"]
description = "Backend for artist-centered image hosting."
name = "acih-backend"
readme = "README.md"
version = "0.1.0"


[tool.poetry.dependencies]
python = "3.11.*"


[tool.poetry.group.app]
optional = true


[tool.poetry.group.app.dependencies]
alembic = {extras = ["tz"], version = "1.12.0"}
fastapi = {version = "0.109.0", extras = ["all"]}
minio = "7.1.17"
psycopg2-binary = "2.9.7"
pydantic = "2.5.3"
pyjwt = "2.8.0"
sqlalchemy = "2.0.20"


[tool.poetry.group.lint]
optional = true


[tool.poetry.group.lint.dependencies]
flake8 = "6.0.0"
mypy = "1.3.0"
pyenchant = "3.2.2"  # necessary for pylint spell checking
pylint = "2.17.4"
pylint-per-file-ignores = "1.2.1"
pylint-pydantic = "0.2.0"  # A Pylint plugin to help Pylint understand the Pydantic.
ruff = "0.0.272"


[tool.poetry.group.test]
optional = true


[tool.poetry.group.test.dependencies]
dirty-equals = "0.7.1.post0"
pytest = "7.3.2"
pytest-cov = "4.1.0"
pytest-env = "1.0.1"


[tool.pylint-per-file-ignores]
"tests/conftest.py" = """ \
    redefined-outer-name, \
"""
